[SLO] Move slo_transform_template into transform_generators
Co-locate slo_transform_template.ts with its only consumers — all 7
transform generators and the base class. Simplifies imports from
'../../assets/transform_templates/slo_transform_template' to
'./slo_transform_template'. Removes the now-empty assets/ directory.
