+++ x-pack/solutions/observability/plugins/slo/server/services/transform_generators/slo_transform_template.ts
